+++ maze_dataset/dataset/dataset.py
@@ -19,7 +19,7 @@ from muutils.tensor_utils import DTYPE_MAP
 from torch.utils.data import Dataset
 from zanj import ZANJ
 
-from maze_dataset.utils.utils import DEFAULT_SEED, GLOBAL_SEED, set_reproducibility
+from muutils.mlutils import DEFAULT_SEED, GLOBAL_SEED, set_reproducibility
 
 
 def _dtype_serialization_fn(datatype: torch.dtype | np.dtype) -> str:
